opt:rewrite wordfind search
